refactor: organise tests according to dependencies
- organises tests according to their dependencies.
- removes unnecessary autouse fixtures
- removes gated imports at the module level
- uses `--ignore` to exclude redis/sqlalchemy/sentry tests from no-extras tests
