+++ tests/conftest.py
@@ -10,6 +10,8 @@ from uuid import uuid4
 import pytest
 from asyncpg.pgproto import pgproto
 
+from starlite_saqlalchemy import constants
+
 if TYPE_CHECKING:
     from collections.abc import Callable
     from pathlib import Path
@@ -23,6 +25,12 @@ if TYPE_CHECKING:
 # Ensure that pytest_dotenv is loaded before
 # so pytest_starlite_saqlalchemy uses correct env values
 pytest_plugins = ("pytest_dotenv", "pytest_starlite_saqlalchemy.plugin")
+
+
+@pytest.fixture()
+def _sqlalchemy_installed() -> None:
+    if not constants.IS_SQLALCHEMY_INSTALLED:
+        pytest.skip("sqlalchemy not installed")
 
 
 @pytest.fixture(name="raw_authors")
@@ -48,7 +56,9 @@ def fx_raw_authors() -> list[dict[str, Any]]:
 
 
 @pytest.fixture(name="authors")
-def fx_authors(raw_authors: list[dict[str, Any]]) -> list[authors.Author]:
+def fx_authors(
+    raw_authors: list[dict[str, Any]], _sqlalchemy_installed: None
+) -> list[authors.Author]:
     """Collection of parsed Author models."""
     from tests.utils.domain import authors
 
@@ -75,7 +85,7 @@ def fx_raw_books(raw_authors: list[dict[str, Any]]) -> list[dict[str, Any]]:
 
 
 @pytest.fixture(name="books")
-def fx_books(raw_books: list[dict[str, Any]]) -> list[books.Book]:
+def fx_books(raw_books: list[dict[str, Any]], _sqlalchemy_installed: None) -> list[books.Book]:
     """Collection of parsed Book models."""
     from tests.utils.domain import books
