+++ tests/unit/redis/test_cache.py
@@ -1,9 +1,5 @@
 """Test for the application cache configurations."""
-# pylint: disable=wrong-import-position
 import pytest
-
-pytest.importorskip("redis")
-
 from starlite.config.cache import default_cache_key_builder
 from starlite.testing import RequestFactory
