+++ tests/unit/sqlalchemy/test_db.py
@@ -1,13 +1,10 @@
 """Tests for db module."""
 # pylint: disable=protected-access
-# pylint: disable=wrong-import-position,wrong-import-order
 from __future__ import annotations
 
-import pytest
-
-pytest.importorskip("sqlalchemy")
-
 from uuid import uuid4
+
+import pytest
 
 from starlite_saqlalchemy import db
