+++ tests/unit/test_testing.py
@@ -2,14 +2,11 @@
 # pylint: disable=wrong-import-position,wrong-import-order
 from __future__ import annotations
 
-import pytest
-
-pytest.importorskip("sqlalchemy")
-
 from typing import TYPE_CHECKING
 from unittest.mock import MagicMock
 
 import httpx
+import pytest
 from starlite.status_codes import (
     HTTP_200_OK,
     HTTP_404_NOT_FOUND,
@@ -17,11 +14,6 @@ from starlite.status_codes import (
 )
 
 from starlite_saqlalchemy import testing
-from starlite_saqlalchemy.exceptions import ConflictError, StarliteSaqlalchemyError
-from starlite_saqlalchemy.testing.generic_mock_repository import GenericMockRepository
-from tests.utils.domain.authors import Author
-from tests.utils.domain.authors import Service as AuthorService
-from tests.utils.domain.books import Book
 
 if TYPE_CHECKING:
     from typing import Any
@@ -29,65 +21,8 @@ if TYPE_CHECKING:
     from pytest import MonkeyPatch
     from starlite import TestClient
 
-
-async def test_repo_raises_conflict_if_add_with_id(
-    authors: list[Author],
-    author_repository: GenericMockRepository[Author],
-) -> None:
-    """Test mock repo raises conflict if add identified entity."""
-    with pytest.raises(ConflictError):
-        await author_repository.add(authors[0])
-
-
-def test_generic_mock_repository_parametrization() -> None:
-    """Test that the mock repository handles multiple types."""
-    author_repo = GenericMockRepository[Author]
-    book_repo = GenericMockRepository[Book]
-    assert author_repo.model_type is Author  # type:ignore[misc]
-    assert book_repo.model_type is Book  # type:ignore[misc]
-
-
-def test_generic_mock_repository_seed_collection(
-    author_repository_type: type[GenericMockRepository[Author]],
-) -> None:
-    """Test seeding instances."""
-    author_repository_type.seed_collection([Author(id="abc")])
-    assert "abc" in author_repository_type.collection
-
-
-def test_generic_mock_repository_clear_collection(
-    author_repository_type: type[GenericMockRepository[Author]],
-) -> None:
-    """Test clearing collection for type."""
-    author_repository_type.clear_collection()
-    assert not author_repository_type.collection
-
-
-def test_generic_mock_repository_filter_collection_by_kwargs(
-    author_repository: GenericMockRepository[Author],
-) -> None:
-    """Test filtering the repository collection by kwargs."""
-    author_repository.filter_collection_by_kwargs(name="Leo Tolstoy")
-    assert len(author_repository.collection) == 1
-    assert list(author_repository.collection.values())[0].name == "Leo Tolstoy"
-
-
-def test_generic_mock_repository_filter_collection_by_kwargs_and_semantics(
-    author_repository: GenericMockRepository[Author],
-) -> None:
-    """Test that filtering by kwargs has `AND` semantics when multiple kwargs,
-    not `OR`."""
-    author_repository.filter_collection_by_kwargs(name="Agatha Christie", dob="1828-09-09")
-    assert len(author_repository.collection) == 0
-
-
-def test_generic_mock_repository_raises_repository_exception_if_named_attribute_doesnt_exist(
-    author_repository: GenericMockRepository[Author],
-) -> None:
-    """Test that a repo exception is raised if a named attribute doesn't
-    exist."""
-    with pytest.raises(StarliteSaqlalchemyError):
-        author_repository.filter_collection_by_kwargs(cricket="ball")
+    from tests.utils.domain.authors import Author
+    from tests.utils.domain.authors import Service as AuthorService
 
 
 @pytest.fixture(name="mock_response")
@@ -113,6 +48,7 @@ def fx_mock_client(
 
 @pytest.fixture(name="tester")
 def fx_tester(
+    author_service_type: type[AuthorService],
     authors: list[Author],
     raw_authors: list[dict[str, Any]],
     mock_client: TestClient,
@@ -126,7 +62,7 @@ def fx_tester(
         base_path="/authors",
         collection=authors[:1],
         raw_collection=raw_authors[:1],
-        service_type=AuthorService,
+        service_type=author_service_type,
         monkeypatch=monkeypatch,
     )
 
@@ -135,10 +71,12 @@ async def test_tester_get_collection_request_service_method_patch(
     tester: testing.ControllerTest, mock_response: MagicMock
 ) -> None:
     """Test that the "list" service method has been patched."""
+    from tests.utils.domain.authors import Service
+
     mock_response.json.return_value = tester.raw_collection
     tester.test_get_collection()
-    assert "<locals>._list" in str(AuthorService.list)
-    assert await AuthorService(session=None).list() == tester.collection
+    assert "<locals>._list" in str(Service.list)
+    assert await Service(session=None).list() == tester.collection
 
 
 def test_tester_get_collection_raises_assertion_error_on_status_code(
@@ -190,11 +128,14 @@ def test_tester_json_in_request_kwargs(
     assert "json" in call.kwargs
 
 
+@pytest.mark.usefixtures("_sqlalchemy_installed")
 async def test_tester_member_request_service_method_patch(tester: testing.ControllerTest) -> None:
     """Test that the appropriate service method gets patched."""
+    from tests.utils.domain.authors import Service
+
     tester.test_member_request("GET", "get", 200)
-    assert "<locals>._method" in str(AuthorService.get)
-    assert await AuthorService(session=None).get(123) == tester.collection[0]
+    assert "<locals>._method" in str(Service.get)
+    assert await Service(session=None).get(123) == tester.collection[0]
 
 
 @pytest.mark.parametrize("params", [{"a": "b"}, None])
