+++ tox.ini
@@ -64,7 +64,7 @@ commands = pytest tests/integration {posargs}
 
 [testenv:no-extras]
 basepython = python3.11
-commands = coverage run -p -m pytest {posargs}
+commands = coverage run -p -m pytest --ignore=tests/unit/redis --ignore=tests/unit/sentry --ignore=tests/unit/sqlalchemy {posargs}
 
 [testenv:docs]
 basepython = python3.11
